Move `babel-cli` dependency to dev-only
This is only used as part of the `prepublish` step, to create a bundled copy
before pushing to npm. No reason to have this as a production dependency that I
can see - it adds more than 1200 lines (!!) of shrinkwrap dependencies to any
projects that depend on this, which translates to minutes of extra install time.