fix: Fixed bug when using babel-register
This plugin is trying to require the webpack config specified in the conf but when a project uses
babel-register require-hook, the plugin is also going to be run on the required webpack config,
which would required the webpack config (itself), etc... and eventually messing up with
require.cache or have an infinite loop. This commit fixes that.

Fixes #19